Web: Moved Location of AuthenticatedBlueprint to REST; Fix #5822
