+++ lib/rucio/web/rest/flaskapi/v1/accountlimits.py
@@ -20,7 +20,7 @@ from rucio.api.account_limit import set_local_account_limit, delete_local_accoun
 from rucio.common.exception import RSENotFound, AccessDenied, AccountNotFound
 from rucio.web.rest.flaskapi.v1.common import response_headers, ErrorHandlingMethodView, \
     generate_http_error_flask, json_parameters, param_get
-from rucio.web.ui.flask.bp import AuthenticatedBlueprint
+from rucio.web.rest.flaskapi.authenticated_bp import AuthenticatedBlueprint
 
 
 class LocalAccountLimit(ErrorHandlingMethodView):
