+++ lib/rucio/web/rest/flaskapi/v1/dirac.py
@@ -21,7 +21,7 @@ from rucio.common.exception import AccessDenied, DataIdentifierAlreadyExists, Da
 from rucio.common.utils import parse_response
 from rucio.web.rest.flaskapi.v1.common import response_headers, generate_http_error_flask, \
     ErrorHandlingMethodView, json_parameters, param_get
-from rucio.web.ui.flask.bp import AuthenticatedBlueprint
+from rucio.web.rest.flaskapi.authenticated_bp import AuthenticatedBlueprint
 
 
 class AddFiles(ErrorHandlingMethodView):
