+++ lib/rucio/web/rest/flaskapi/v1/heartbeats.py
@@ -18,7 +18,7 @@ import json
 from flask import Flask, Response, request
 
 from rucio.api.heartbeat import list_heartbeats, create_heartbeat
-from rucio.web.ui.flask.bp import AuthenticatedBlueprint
+from rucio.web.rest.flaskapi.authenticated_bp import AuthenticatedBlueprint
 from rucio.common.utils import APIEncoder
 from rucio.common.exception import UnsupportedValueType, UnsupportedKeyType, KeyNotFound, AccessDenied
 from rucio.web.rest.flaskapi.v1.common import response_headers, check_accept_header_wrapper_flask, \
