+++ lib/rucio/web/rest/flaskapi/v1/rses.py
@@ -29,7 +29,7 @@ from rucio.common.utils import Availability, render_json, APIEncoder
 from rucio.rse import rsemanager
 from rucio.web.rest.flaskapi.v1.common import response_headers, check_accept_header_wrapper_flask, \
     try_stream, generate_http_error_flask, ErrorHandlingMethodView, json_parameters, param_get
-from rucio.web.ui.flask.bp import AuthenticatedBlueprint
+from rucio.web.rest.flaskapi.authenticated_bp import AuthenticatedBlueprint
 
 
 class RSEs(ErrorHandlingMethodView):
