+++ lib/rucio/web/rest/flaskapi/v1/rules.py
@@ -28,7 +28,7 @@ from rucio.common.exception import InsufficientAccountLimit, RuleNotFound, Acces
 from rucio.common.utils import render_json, APIEncoder
 from rucio.web.rest.flaskapi.v1.common import check_accept_header_wrapper_flask, parse_scope_name, try_stream, \
     response_headers, generate_http_error_flask, ErrorHandlingMethodView, json_parameters, param_get
-from rucio.web.ui.flask.bp import AuthenticatedBlueprint
+from rucio.web.rest.flaskapi.authenticated_bp import AuthenticatedBlueprint
 
 
 class Rule(ErrorHandlingMethodView):
