+++ lib/rucio/web/rest/flaskapi/v1/subscriptions.py
@@ -25,7 +25,7 @@ from rucio.common.exception import InvalidObject, SubscriptionDuplicate, Subscri
 from rucio.common.utils import render_json, APIEncoder
 from rucio.web.rest.flaskapi.v1.common import check_accept_header_wrapper_flask, try_stream, \
     response_headers, generate_http_error_flask, ErrorHandlingMethodView, json_parameters, param_get
-from rucio.web.ui.flask.bp import AuthenticatedBlueprint
+from rucio.web.rest.flaskapi.authenticated_bp import AuthenticatedBlueprint
 
 
 class Subscription(ErrorHandlingMethodView):
